+++ changelog.md
@@ -15,3 +15,8 @@
 ### v0.1.3
 - November 5, 2024
 - It can fetch API nad loads the information after searching (Updating main.js)
+
+### v0.1.4
+- November 6, 2024
+- Card Adjustments on Search Results
+- It can now search both Anime and Rating Rx Anime (HAnime)